+++ gen/com/intellij/idea/plugin/hybris/impex/psi/impl/ImpexValueGroupImpl.java
@@ -52,10 +52,4 @@ public class ImpexValueGroupImpl extends ImpexValueGroupMixin implements ImpexVa
     return findChildByClass(ImpexValue.class);
   }
 
-  @Override
-  @Nullable
-  public String computeValue() {
-    return ImpexPsiUtil.computeValue(this);
-  }
-
 }