+++ src/com/intellij/idea/plugin/hybris/impex/inspection/analyzer/ImpexModificationValueAnalyzer.kt
@@ -1,6 +1,6 @@
 /*
- * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
+ * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -21,8 +21,10 @@ package com.intellij.idea.plugin.hybris.impex.inspection.analyzer
 import com.intellij.codeInspection.ProblemHighlightType
 import com.intellij.codeInspection.ProblemsHolder
 import com.intellij.idea.plugin.hybris.impex.constants.modifier.AttributeModifier
-import com.intellij.idea.plugin.hybris.impex.psi.*
-import com.intellij.idea.plugin.hybris.impex.utils.ImpexPsiUtils
+import com.intellij.idea.plugin.hybris.impex.psi.ImpexFile
+import com.intellij.idea.plugin.hybris.impex.psi.ImpexFullHeaderParameter
+import com.intellij.idea.plugin.hybris.impex.psi.ImpexHeaderLine
+import com.intellij.idea.plugin.hybris.impex.psi.ImpexTypes
 import com.intellij.idea.plugin.hybris.psi.util.PsiTreeUtilExt
 import com.intellij.psi.PsiElement
 import com.intellij.psi.util.PsiTreeUtil
@@ -129,9 +131,11 @@ class DataTable(private val keyRows: List<Key>, private val attrs: List<String>,
             val row = Row(keyValue, bitSet, arrayOfNulls(attrs.size))
             attrs.forEach { av ->
                 val valueGroups =
-                    attrsValues.filter { it.text == av }
+                    attrsValues
+                        .asSequence()
+                        .filter { it.text == av }
                         .filter { hasNoAppendModeModifier(it) }
-                        .flatMap { ImpexPsiUtils.getColumnForHeader(it) }
+                        .flatMap { it.valueGroups }
                         .filter { it.value != null }
                         .filter {
                             PsiTreeUtilExt.getLeafsOfAnyElementType(
@@ -145,6 +149,7 @@ class DataTable(private val keyRows: List<Key>, private val attrs: List<String>,
                             val commonContext = PsiTreeUtil.findCommonContext(keyValue.keys.first(), it)
                             commonContext != null && commonContext !is ImpexFile
                         }
+                        .toList()
 
                 if (valueGroups.isNotEmpty()) {
                     valueGroups.forEach { valueGroup ->
@@ -155,14 +160,14 @@ class DataTable(private val keyRows: List<Key>, private val attrs: List<String>,
                                 row.valueGroup[indexOf] = null
                             }
                         } else {
-                            val headerForValueGroup = ImpexPsiUtils.getHeaderForValueGroup(valueGroup as? ImpexValueGroup)
-                            if (headerForValueGroup is ImpexFullHeaderParameter) {
-                                val indexOf = attrs.indexOfFirst { headerForValueGroup.text == it }
-                                if (indexOf > -1) {
-                                    bitSet[indexOf] = 1
-                                    row.valueGroup[indexOf] = valueGroup
+                            valueGroup.fullHeaderParameter
+                                ?.let { headerParameter ->
+                                    val indexOf = attrs.indexOfFirst { headerParameter.text == it }
+                                    if (indexOf > -1) {
+                                        bitSet[indexOf] = 1
+                                        row.valueGroup[indexOf] = valueGroup
+                                    }
                                 }
-                            }
                         }
                     }
                 }