+++ src/com/intellij/idea/plugin/hybris/impex/psi/impl/ImpexValueGroupMixin.kt
@@ -19,12 +19,14 @@
 package com.intellij.idea.plugin.hybris.impex.psi.impl
 
 import com.intellij.extapi.psi.ASTWrapperPsiElement
+import com.intellij.idea.plugin.hybris.impex.constants.modifier.AttributeModifier
 import com.intellij.idea.plugin.hybris.impex.psi.ImpexFullHeaderParameter
 import com.intellij.idea.plugin.hybris.impex.psi.ImpexValueGroup
 import com.intellij.idea.plugin.hybris.impex.psi.ImpexValueLine
 import com.intellij.idea.plugin.hybris.impex.utils.ImpexPsiUtils
 import com.intellij.lang.ASTNode
 import com.intellij.openapi.util.Key
+import com.intellij.openapi.util.text.StringUtil
 import com.intellij.psi.util.*
 import java.io.Serial
 
@@ -41,8 +43,9 @@ abstract class ImpexValueGroupMixin(node: ASTNode) : ASTWrapperPsiElement(node),
     }, false)
 
     override fun getFullHeaderParameter(): ImpexFullHeaderParameter? = CachedValuesManager.getManager(project).getCachedValue(this, CACHE_KEY_FULL_HEADER_PARAMETER, {
-        val header = ImpexPsiUtils
-            .getHeaderForValueGroup(this) as? ImpexFullHeaderParameter
+        val header = this.valueLine
+            ?.headerLine
+            ?.getFullHeaderParameter(this.columnNumber)
 
         CachedValueProvider.Result.createSingleDependency(
             header,
@@ -59,10 +62,34 @@ abstract class ImpexValueGroupMixin(node: ASTNode) : ASTWrapperPsiElement(node),
         )
     }, false)
 
+    override fun computeValue(): String? = CachedValuesManager.getManager(project).getCachedValue(this, CACHE_KEY_VALUE_OR_DEFAULT, {
+        val computedValue = this
+            .value
+            ?.text
+            ?: this.fullHeaderParameter
+                ?.getAttribute(AttributeModifier.DEFAULT)
+                ?.anyAttributeValue
+                ?.let {
+                    it.stringList.firstOrNull()
+                        ?.text
+                        ?: it.text
+                }
+
+        val defaultValue = computedValue
+            ?.let { StringUtil.unquoteString(it) }
+            ?.trim()
+
+        CachedValueProvider.Result.createSingleDependency(
+            defaultValue,
+            PsiModificationTracker.MODIFICATION_COUNT,
+        )
+    }, false)
+
     companion object {
         val CACHE_KEY_VALUE_LINE = Key.create<CachedValue<ImpexValueLine?>>("SAP_CX_IMPEX_VALUE_LINE")
         val CACHE_KEY_FULL_HEADER_PARAMETER = Key.create<CachedValue<ImpexFullHeaderParameter?>>("SAP_CX_IMPEX_FULL_HEADER_PARAMETER")
         val CACHE_KEY_COLUMN_NUMBER = Key.create<CachedValue<Int>>("SAP_CX_IMPEX_COLUMN_NUMBER")
+        val CACHE_KEY_VALUE_OR_DEFAULT = Key.create<CachedValue<String>>("SAP_CX_IMPEX_VALUE_OR_DEFAULT")
 
         @Serial
         private val serialVersionUID: Long = -4491471414641409161L