+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Migrated `Remove Column` to coroutine with progress [#1169](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1169)
 - Migrated `Remove Table` to coroutine with progress [#1170](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1170)
 - Migrated `Split Table` to coroutine with progress [#1171](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1171)
+- Re-enabled `odd` | `even` value lines coloring [#1178](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1178)
 
 ### `Cockpit NG` enhancements
 - Added `Spring EL` language injection into the body of the `y:shortLabel` tag [#1168](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1168)